refactor(remote-config): migrate rcservicemrf to v2 component
