+++ cmd/agent/subcommands/run/command.go
@@ -159,7 +159,7 @@ import (
 	remoteconfig "github.com/DataDog/datadog-agent/comp/remote-config"
 	"github.com/DataDog/datadog-agent/comp/remote-config/rcclient"
 	rcservicefx "github.com/DataDog/datadog-agent/comp/remote-config/rcservice/fx"
-	"github.com/DataDog/datadog-agent/comp/remote-config/rcservicemrf/rcservicemrfimpl"
+	rcservicemrffx "github.com/DataDog/datadog-agent/comp/remote-config/rcservicemrf/fx"
 	rctelemetryreporterfx "github.com/DataDog/datadog-agent/comp/remote-config/rctelemetryreporter/fx"
 	metricscompressorfx "github.com/DataDog/datadog-agent/comp/serializer/metricscompression/fx"
 	snmpscanmanager "github.com/DataDog/datadog-agent/comp/snmpscanmanager/def"
@@ -473,7 +473,7 @@ func getSharedFxOption() fx.Option {
 		hostProfilerFlareFx.Module(),
 		rctelemetryreporterfx.Module(),
 		rcservicefx.Module(),
-		rcservicemrfimpl.Module(),
+		rcservicemrffx.Module(),
 		remoteconfig.Bundle(),
 		daemoncheckerfx.Module(),
 		fleetfx.Module(),
