+++ comp/api/grpcserver/impl-agent/grpc.go
@@ -31,7 +31,7 @@ import (
 	replay "github.com/DataDog/datadog-agent/comp/dogstatsd/replay/def"
 	dogstatsdServer "github.com/DataDog/datadog-agent/comp/dogstatsd/server"
 	rcservice "github.com/DataDog/datadog-agent/comp/remote-config/rcservice/def"
-	"github.com/DataDog/datadog-agent/comp/remote-config/rcservicemrf"
+	rcservicemrf "github.com/DataDog/datadog-agent/comp/remote-config/rcservicemrf/def"
 	pb "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
 	grpcutil "github.com/DataDog/datadog-agent/pkg/util/grpc"
 	"github.com/DataDog/datadog-agent/pkg/util/option"
