+++ comp/remote-config/rcservicemrf/def/component.go
